requirements.txt
fastapi
uvicorn
aiofiles
python-dotenv